% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{chm_add_modality}
\alias{chm_add_modality}
\title{Register a modality encoder}
\usage{
chm_add_modality(task, name, encoder)
}
\arguments{
\item{task}{chm_task}

\item{name}{modality name present in task$modalities}

\item{encoder}{a function taking (se, id_col) and returning a feature matrix/data.frame with rownames as sample ids}
}
\description{
Register a modality encoder
}
