% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochlear_map.R
\name{cochlear_map}
\alias{cochlear_map}
\title{cochlear_map}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
cochlear_map(f, Fmax = 20000)
}
\arguments{
\item{f}{Frequency in Hz}

\item{Fmax}{Maximum frequency represented on the basilar membrane in Hz. By default, this is
20 kHz, the value for the human cochlea. Setting Fmax to 57 kHz gives a map appropriate for the cat
cochlea.}
}
\value{
Returns percent distance from apex of basilar membrane

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{cochlear_map} Convert frequency to distance along the basilar membrane using M.C. Liberman's
cochlear frequency map for the cat.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
inv_cochlear_map
}
