% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_xbm_bands.R
\name{equal_xbm_bands}
\alias{equal_xbm_bands}
\title{equal_xbm_bands}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
equal_xbm_bands(fmin, fmax, N)
}
\arguments{
\item{fmin}{minimum frequency in Hz}

\item{fmax}{maximum frequency in Hz}

\item{N}{number of frequency bands}
}
\value{
Returns a vector of band cutoff frequencies in Hz

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{equal_xbm_bands} Divide frequency interval into N bands of equal width along
the human basilar membrane. Based on M.C. Liberman's cochlear frequency map
for the cat scaled to match human frequency range of hearing.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
inv_cochlear_map
}
