% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_envelope.R
\name{extract_envelope}
\alias{extract_envelope}
\title{extract_envelope}
\source{
Method based on MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
extract_envelope(x, f = 1, cutoffs, complex = FALSE)
}
\arguments{
\item{x}{signal}

\item{f}{sampling rate in Hz}

\item{cutoffs}{band cutoff frequencies}

\item{complex}{if TRUE returns envelope and fine stucture as a complex valued vector}
}
\value{
a list of amplitude envelopes for the specified cutoffs, or a vector if only two cutoffs are provided
}
\description{
\code{extract_envelope} Extract amplitude envelope of a signal.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
make_band_chimeras
}
