% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{grf_fit}
\alias{grf_fit}
\title{Fit a multimodal model}
\usage{
grf_fit(
  task,
  learner,
  outcome_name,
  family = c("classification", "regression")
)
}
\arguments{
\item{learner}{a parsnip model spec}

\item{outcome_name}{column name in outcome table}

\item{family}{'classification'|'regression'}
}
\description{
Fit a multimodal model
}
