% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{grf_fuse}
\alias{grf_fuse}
\title{Fuse encoded modalities into meta-features}
\usage{
grf_fuse(task, mode = c("early", "late", "hybrid"))
}
\arguments{
\item{mode}{one of 'early','late','hybrid'}
}
\description{
Fuse encoded modalities into meta-features
}
