% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{grf_task}
\alias{grf_task}
\title{Create a multimodal task}
\usage{
grf_task(mae, outcome, id_col = "sample_id")
}
\arguments{
\item{mae}{MultiAssayExperiment or list of modality tables. For MVP we accept
either; lists will be coerced to MAE with synthetic assays.}

\item{outcome}{A tibble/data.frame with one column outcome (factor/numeric) and optional id column name \code{sample_id}.}

\item{id_col}{Name of sample id column shared across modalities (default 'sample_id').}
}
\value{
An object of class \code{grf_task}.
}
\description{
Create a multimodal task
}
