% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_cochlear_map.R
\name{inv_cochlear_map}
\alias{inv_cochlear_map}
\title{inv_cochlear_map}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
inv_cochlear_map(x, Fmax = 20000)
}
\arguments{
\item{x}{Percent distance from apex of basilar membrane}

\item{Fmax}{Maximum frequency represented on the basilar membrane in Hz. By default, this is
57 kHz, the value for the cat. Setting Fmax to 20 kHz gives a map appropriate for the human
cochlea.}
}
\value{
Returns frequency in Hz

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{inv_cochlear_map} Convert distance along the basilar membrane to frequency using M.C. Liberman's
cochlear frequency map for the cat.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
cochlear_map
}
