% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_band_chimeras.R
\name{make_band_chimeras}
\alias{make_band_chimeras}
\title{make_band_chimeras}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
make_band_chimeras(
  x1,
  x2 = "noise",
  f = 1,
  n_bands = c(1, 8, 32),
  savepath = ".",
  bitrate = 16,
  refilter = FALSE,
  save_waves = TRUE,
  do_plot = FALSE
)
}
\arguments{
\item{x1, x2}{original signals. Can be a numeric vector or object of class "Wave"}

\item{f}{sampling rate in Hz. Not required if using a "Wave" object. Must match for both signals.}

\item{n_bands}{Number of frequency bands used for synthesis. Can be scalar or vector, in which case one pair of
chimeras is synthesized for each element of n_bands.}

\item{savepath}{path to folder where chimeras will be saved}

\item{bitrate}{bitrate of signal. 16 used as default, if "Wave" object is used bitrate will be extracted from x1}

\item{refilter}{set to TRUE to filter again after exchange operation (default FALSE). Not tested.}

\item{save_waves}{if FALSE waves will be returned in a list}

\item{do_plot}{plot spectrograms of chimeras. Not implremented.}
}
\value{
wav files of chimera pairs for each band in n_bands

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{make_band_chimeras} Synthesize and store single or multi-band auditory chimeras from pair of original signals.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
multi_band_chimera equal_xbm_bands
}
