% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_band_chimera.R
\name{multi_band_chimera}
\alias{multi_band_chimera}
\title{multi_band_chimera}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
multi_band_chimera(x1, x2, f = 1, cutoffs, fb = NULL, refilter = 0)
}
\arguments{
\item{x1, x2}{original signals}

\item{f}{sampling rate in Hz}

\item{cutoffs}{band cutoff frequencies}

\item{fb}{filter banks. Not required if cutoffs are specified}

\item{refilter}{set to 1 to filter again after exchange operation (default 0)}
}
\value{
a list of two auditory chimeras. The first has envelope of x1 and the
fine structure of x2 and the second has the envelope of x2 and the fine structure
of x1

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{multi_band_chimera} Synthesize pair of multi-band "auditory chimeras" by
dividing each signal into frequency bands, then interchanging envelope and
fine structure in each band using Hilbert transforms.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
quad_filt_bank
}
