% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_matched_noise.R
\name{psd_matched_noise}
\alias{psd_matched_noise}
\title{psd_matched_noise}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
psd_matched_noise(x)
}
\arguments{
\item{x}{A vector of a sound wave or an object of class "Wave"}
}
\value{
Returns a vector of equal length or an object of class "Wave".

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{psd_matched_noise} Synthesize noise N having the same power spectrum and duration as original signal
X. Done by randomizing phase of Fourier spectrum.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
