% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_filt_bank.R
\name{quad_filt_bank}
\alias{quad_filt_bank}
\title{quad_filt_bank}
\source{
Translated from MATLAB code in Smith, Delgutte & Oxenham (2002).
}
\usage{
quad_filt_bank(cutoffs, f = 1, N = NULL)
}
\arguments{
\item{cutoffs}{band cutoff frequencies}

\item{f}{sampling rate. 1 will be used if sample rate is not provided.}

\item{N}{filter order. Default is 8*f/(min(bandwidth))}
}
\value{
Returns a list of filter banks

Copyright Bertrand Delgutte, 1999-2000
}
\description{
\code{quad_filt_bank} Create bank of FIR complex filters whose real and imaginary parts are in quadrature.
}
\references{
Smith, Z. M., Delgutte, B., & Oxenham, A. J. (2002). Chimaeric sounds reveal dichotomies in auditory perception. Nature, 416(6876), 87–90. doi:10.1038/416087a
}
\seealso{
equal_xbm_bands
}
